# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialCommander API - Social media management for X and Reddit
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts with optional filters
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, published, failed, all]
        - name: platform
          in: query
          schema:
            type: string
            enum: [twitter, reddit, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Immediately publish a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/schedule:
    post:
      operationId: schedulePost
      tags: [posts]
      summary: Schedule a post for later
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "200":
          description: Scheduled post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/calendar:
    get:
      operationId: getCalendar
      tags: [posts]
      summary: Get calendar view of scheduled posts
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Calendar data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"
  /posts/recent:
    get:
      operationId: getRecentPosts
      tags: [posts]
      summary: Get recently published posts with engagement data
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, platform, username, displayName, color, status, createdAt]
      properties:
        id:
          type: integer
        platform:
          type: string
          enum: [twitter, reddit]
        username:
          type: string
        displayName:
          type: string
        color:
          type: string
          description: Hex color code for UI identification
        status:
          type: string
          enum: [active, suspended, paused]
        avatarUrl:
          type: ["string", "null"]
        proxyConfig:
          type: ["string", "null"]
          description: Proxy URL for this account
        voiceProfile:
          type: ["string", "null"]
          description: AI voice/tone profile description
        postsCount:
          type: integer
          default: 0
        followersCount:
          type: integer
          default: 0
        engagementRate:
          type: number
          default: 0
        lastPostAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Post:
      type: object
      required: [id, accountId, platform, content, status, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        platform:
          type: string
          enum: [twitter, reddit]
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, scheduled, published, failed]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        publishedAt:
          type: ["string", "null"]
          format: date-time
        externalId:
          type: ["string", "null"]
        impressions:
          type: integer
          default: 0
        likes:
          type: integer
          default: 0
        comments:
          type: integer
          default: 0
        reposts:
          type: integer
          default: 0
        aiGenerated:
          type: boolean
          default: false
        subreddit:
          type: ["string", "null"]
        postTitle:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        retryCount:
          type: integer
          default: 0
        account:
          $ref: "#/components/schemas/Account"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [accountId, content]
      properties:
        accountId:
          type: integer
        content:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, scheduled]
          default: draft
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        subreddit:
          type: ["string", "null"]
        postTitle:
          type: ["string", "null"]
        aiGenerated:
          type: boolean
          default: false
    PostUpdate:
      type: object
      properties:
        content:
          type: ["string", "null"]
        mediaUrls:
          type: array
          items:
            type: string
        status:
          type: ["string", "null"]
          enum: [draft, scheduled, published, failed]
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        subreddit:
          type: ["string", "null"]
        postTitle:
          type: ["string", "null"]
    PostsPage:
      type: object
      required: [posts, total, limit, offset]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ScheduleInput:
      type: object
      required: [scheduledAt]
      properties:
        scheduledAt:
          type: string
          format: date-time
        jitterMinutes:
          type: integer
          default: 0
          description: Add random jitter for human-like behavior
    CalendarEntry:
      type: object
      required: [date, posts]
      properties:
        date:
          type: string
          format: date
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"

    # ── ANALYTICS SCHEMAS ─────────────────────────────────────────────────
...