# lib/api-zod/src/generated/types/calendarEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SocialCommander API - Social media management for X and Reddit
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface CalendarEntry {
  date: Date;
  posts: Post[];
}