# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SocialCommander API - Social media management for X and Reddit
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleInput {
  scheduledAt: Date;
  /** Add random jitter for human-like behavior */
  jitterMinutes?: number;
}